azure_function_sample/requirements.txt
azure-functions==1.3.0
azureml-sdk==1.12.0
opencensus-ext-azure==1.0.4
flake8==3.8.3
pytest==6.0.1
mock==4.0.2